Encrypted values get re-encrypted
So, it seems and most definitely is that You the bundle simply encrypts the value no matter if it was encrypted or not. I found and issue when I did this:
- Project `A`:
   -  contain some data,
   - data is encrypted before sending to api,
   - api call is done to project `B`,
- Project `B`
   - runs into timeout (clearly php_ini stuff, but that's not the point),
   - partially imported data is not working

So the timeout happens because the `B` encrypts tones of the data again.
Imported data is not working since the already encrypted data get's re-encrypted.